fishing pole on c up

+++ soh/include/functions.h
@@ -1078,7 +1078,7 @@ bool Inventory_HatchPocketCucco(PlayState* play);
 void Interface_SetDoAction(PlayState* play, u16 action);
 void Interface_LoadActionLabelB(PlayState* play, u16 action);
 s32 Health_ChangeBy(PlayState* play, s16 healthChange);
-void Rupees_ChangeBy(s16 rupeeChange);
+void Rupees_ChangeBy(s64 rupeeChange);
 void Inventory_ChangeAmmo(s16 item, s16 ammoChange);
 void Magic_Fill(PlayState* play);
 void Magic_Reset(PlayState* play);
